ValueCheck.java: renamed to Checker   now check() throws CheckException which must be catched
ValueChecks.java: renamed to Checks
  add static check methods. e.g. setOnlyOne(args)

BasicCLI.java:
  _version/_help(out) don't throw out ControlBreak, but the (void) version does.

Memory.java: offset(long)

MemoryOffset.java: renamed to MemoryWrapOffset

ArrayMemory.java: merge offsets.

types/*: never throw AccessException in some static functions

TypeParsers.java: add Hex/Base64 byte[] parsers.

IntMath.java: add min(int/long) helper func.

MethodParameter.java: remove '$' in default arg-name

+++ bas/mod/Data Types/src/net/bodz/bas/mem/Memory.java
@@ -17,7 +17,7 @@ public interface Memory {
 
     void write(int addr, byte[] buf) throws AccessException;
 
-    Memory offset(int off);
+    Memory offset(long off);
 
     int LITTLE_ENDIAN = 0;
     int BIG_ENDIAN    = 1;
